Update step3.tmpl
Fixed a few typos and mistakes.

Removed 'We need to tell Typescript which files to compile. Those are in the `include` key.' since the include key is not used in the file.
Replaced yarn add -D prettier with yarn add --dev prettier since --dev is used everywhere and explained, while -D is unknown for the newcomer.